+++ BetterUI/PluginInfo.cs
@@ -4,5 +4,5 @@ internal static class PluginInfo
 {
     internal const string Guid = "com.holden.infernoprotocol.betterui";
     internal const string Name = "BetterUI";
-    internal const string Version = "0.9.2";
+    internal const string Version = "0.9.3";
 }